@@ -3,8 +3,8 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
   
-  before_filter :configure_permitted_parameters, if: :devise_controller?
-  before_filter :set_locale
+  before_action :configure_permitted_parameters, if: :devise_controller?
+  before_action :set_locale
   
   def set_locale
     if user_signed_in?
@@ -27,8 +27,8 @@ class ApplicationController < ActionController::Base
   protected
   
   def configure_permitted_parameters
-      devise_parameter_sanitizer.for(:sign_up) << :screenname
-      devise_parameter_sanitizer.for(:account_update) << :screenname    
+      devise_parameter_sanitizer.permit(:sign_up, keys: [:screenname])
+      devise_parameter_sanitizer.permit(:account_update, keys: [:screenname]) 
   end
   
   
