@@ -1,4 +1,4 @@
-class CreateImages < ActiveRecord::Migration
+class CreateImages < ActiveRecord::Migration[5.0]
   def change
     create_table :images do |t|
       t.binary :image_data
