@@ -1,4 +1,4 @@
-class CreateLikes < ActiveRecord::Migration
+class CreateLikes < ActiveRecord::Migration[5.0]
   def change
     create_table :likes do |t|
       t.references :user, index: true
