@@ -1,4 +1,4 @@
-class CreateFollowers < ActiveRecord::Migration
+class CreateFollowers < ActiveRecord::Migration[5.0]
   def change
     create_table :followers do |t|
       t.references :user_followed, index: true
