@@ -1,4 +1,4 @@
-class AddScreennameToUsers < ActiveRecord::Migration
+class AddScreennameToUsers < ActiveRecord::Migration[5.0]
   def change
 	  add_column :users, :screenname, :string
   end
