@@ -1,4 +1,4 @@
-class AddFilenameToImages < ActiveRecord::Migration
+class AddFilenameToImages < ActiveRecord::Migration[5.0]
   def change
     add_column :images, :file_name, :string
   end
