@@ -1,4 +1,4 @@
-class ChangeDatatypeInUserInfo < ActiveRecord::Migration
+class ChangeDatatypeInUserInfo < ActiveRecord::Migration[5.0]
   def change
 	  remove_column :user_infos, :file_name, :text
 	  remove_column :user_infos, :file_type, :text
