@@ -1,4 +1,4 @@
-class AddLocaleToUser < ActiveRecord::Migration
+class AddLocaleToUser < ActiveRecord::Migration[5.0]
   def change
 	add_column :users, :locale, :string
   end
